% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomial_sigma_G}
\alias{binomial_sigma_G}
\title{Function for deriving the estimated standard deviation of the genetic variants
from the first-order and second-order statistics, assuming these are
binomially distributed.}
\usage{
binomial_sigma_G(SS, n = 2)
}
\arguments{
\item{SS}{Numeric matrix of first-order and second-order statistics.}

\item{n}{Integer number of alleles (trials) for the binomial genetic variables.}
}
\value{
Numeric vector containing the standard deviations of the genetic variants.
}
\description{
Function for deriving the estimated standard deviation of the genetic variants
from the first-order and second-order statistics, assuming these are
binomially distributed.
}
\examples{
J <- 5 # number of instruments
par <- random_Gaussian_parameters(J)
dat <- generate_data_MASSIVE_model(N = 1000, n = 2, p = rep(0.3, J), par)
binomial_sigma_G(dat$SS)
}
