% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BMI_psoriasis_samples}
\alias{BMI_psoriasis_samples}
\title{A dataset containing results for reproducing Figure 14 in Section 4 of the UAI article.}
\format{
An object of class \code{list} of length 3.
}
\usage{
BMI_psoriasis_samples
}
\description{
A dataset containing results for reproducing Figure 14 in Section 4 of the UAI article.
}
\keyword{datasets}
