% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_model_handlers.R
\name{decode_IV_model}
\alias{decode_IV_model}
\title{Function for decoding a character vector representing IV model.}
\usage{
decode_IV_model(code, sd_slab, sd_spike)
}
\arguments{
\item{code}{Character vector encoding IV model.}

\item{sd_slab}{Standard deviation of slab component.}

\item{sd_spike}{Standard deviation of spike component.}
}
\value{
List describing prior for IV model.
}
\description{
Function for decoding a character vector representing IV model.
}
\examples{
decode_IV_model(code = get_random_IV_model(5), sd_slab = 1, sd_spike = 0.01)
}
