% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_hyperparameters}
\alias{determine_hyperparameters}
\title{Functioning for automatically determining MASSIVE hyperparameters from data,
where we assume a N(0, 10) prior on skappa_X.}
\usage{
determine_hyperparameters(J, N, SS, sigma_G)
}
\arguments{
\item{J}{Integer number of candidate instruments.}

\item{N}{Integer number of observations.}

\item{SS}{Numeric matrix containing first- and second-order statistics.}

\item{sigma_G}{Numeric vector of instrument standard deviations.}
}
\value{

}
\description{
Functioning for automatically determining MASSIVE hyperparameters from data,
where we assume a N(0, 10) prior on skappa_X.
}
\examples{
J <- 5 # number of instruments
N <- 1000 # number of samples
par <- random_Gaussian_parameters(J)
dat <- generate_data_MASSIVE_model(N, n = 2, p = rep(0.3, J), par)
determine_hyperparameters(J, N, dat$SS, binomial_sigma_G(dat$SS))
}
