% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry2j.R
\name{dry2j}
\alias{dry2j}
\title{Convert (ash) dry weight to approximate energy content}
\usage{
dry2j(dry)
}
\arguments{
\item{dry}{A vector of dry masses (in kg)}
}
\value{
The equivalent energy in Joules
}
\description{
This performs a simple conversion of ash dry weight to energy via Peters' approximate conversion.
}
\examples{
v <- rnorm(20, 10, 2)
v
dry2j(v)
}
\references{
Peters, R.H. (1983) The Ecological Implications of Body Size. Cambridge: Cambridge University Press (Cambridge Studies in Ecology), pp. 231–296. \link{https://doi.org/10.1017/CBO9780511608551}, Appendix Ia
}
\author{
Francis Windram
}
\concept{converter}
