% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior_optima.R
\name{find_optimum}
\alias{find_optimum}
\title{Routine for finding a MASSIVE posterior local optimum}
\usage{
find_optimum(
  J,
  N,
  SS,
  sigma_G,
  prior_sd,
  skappa_X,
  skappa_Y,
  post_fun = scaled_nl_posterior_log,
  gr_fun = scaled_nl_gradient_log,
  hess_fun = scaled_nl_hessian_log
)
}
\arguments{
\item{J}{Integer number of candidate instruments}

\item{N}{Integer number of observations}

\item{SS}{Numeric moments matrix}

\item{sigma_G}{Numeric vector of instrument variances}

\item{prior_sd}{List containing prior hyperparameters}

\item{skappa_X}{Numeric value of confounding coefficient used for initialization}

\item{skappa_Y}{Numeric value of confounding coefficient used for initialization}

\item{post_fun}{Function computing the negative log-posterior}

\item{gr_fun}{Function computing the negative log-posterior gradient}

\item{hess_fun}{Function computing the negative log-posterior Hessian}
}
\value{
Local posterior optimum
}
\description{
Routine for finding a MASSIVE posterior local optimum
}
