% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{five_optima_posterior_surface}
\alias{five_optima_posterior_surface}
\title{A dataset containing results for reproducing Figure 3 in Section 4 of the UAI article.}
\format{
A list containing all the necessary elements for reproducing the
posterior contour surface in Figure 3.
\describe{
\item{optima}{List containing the parameters found at each of the five
local optima in list format.}
\item{grid_values}{Data frame containing maximum posterior values (z)
computed on a two-dimensional grid (x, y) of values for the scale-free
confounding coefficients (skappa_X, skappa_Y).}
\item{simulation_details}{Details of the data and generating model in
the simulation.}
}
}
\usage{
five_optima_posterior_surface
}
\description{
A dataset containing results for reproducing Figure 3 in Section 4 of the UAI article.
}
\keyword{datasets}
