% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data_MASSIVE_model}
\alias{generate_data_MASSIVE_model}
\title{Function to generate data from a MASSIVE generating model.}
\usage{
generate_data_MASSIVE_model(N, n, p, par, seed = NULL)
}
\arguments{
\item{N}{Integer number of observations.}

\item{n}{Integer number of alleles (trials) for the binomial genetic variables.}

\item{p}{Numeric expected allele frequencies for the genetic variables.}

\item{par}{List of MASSIVE model parameters}

\item{seed}{Integer representing seed for random number generation.}
}
\value{
List containing generate data vector as well as the scatter matrix
of first-order and second-order statistics.
}
\description{
Function to generate data from a MASSIVE generating model.
}
\examples{
J <- 5 # number of instruments
par <- random_Gaussian_parameters(J)
generate_data_MASSIVE_model(N = 1000, n = 2, p = rep(0.3, J), par)
}
