% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_model_handlers.R
\name{get_empty_IV_model}
\alias{get_empty_IV_model}
\title{Function for generating the empty IV model (only spike priors).}
\usage{
get_empty_IV_model(J)
}
\arguments{
\item{J}{Integer number of candidate instruments.}
}
\value{
Character vector representing empty IV model.
}
\description{
Function for generating the empty IV model (only spike priors).
}
\examples{
get_empty_IV_model(5)
}
