% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_model_handlers.R
\name{get_full_IV_model}
\alias{get_full_IV_model}
\title{Function for generating the full IV model (only slab priors).}
\usage{
get_full_IV_model(J)
}
\arguments{
\item{J}{Integer number of candidate instruments.}
}
\value{
Character vector representing full IV model.
}
\description{
Function for generating the full IV model (only slab priors).
}
\examples{
get_full_IV_model(5)
}
