% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_handlers.R
\name{get_ML_solution}
\alias{get_ML_solution}
\title{Function for deriving maximum likelihood solution given values for scale-free
confounding coefficients.}
\usage{
get_ML_solution(SS, skappa_X, skappa_Y, sigma_G)
}
\arguments{
\item{SS}{Numeric matrix containing first- and second-order statistics.}

\item{skappa_X}{Numeric scale-free confounding coefficient on exposure.}

\item{skappa_Y}{Numeric scale-free confounding coefficient on outcome.}

\item{sigma_G}{Numeric vector of instrument standard deviations.}
}
\value{
List of parameters on the ML manifold with the same values for
the scaled confounding coefficients (skappa_X, skappa_Y).
}
\description{
Function for deriving maximum likelihood solution given values for scale-free
confounding coefficients.
}
\examples{
J <- 5 # number of instruments
par <- random_Gaussian_parameters(J)
generate_data_MASSIVE_model(N = 1000, n = 2, p = rep(0.3, J), par)
}
