% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_model_handlers.R
\name{get_random_IV_model}
\alias{get_random_IV_model}
\title{Function for generating a random IV model (combination of slab and spike priors).}
\usage{
get_random_IV_model(J)
}
\arguments{
\item{J}{Integer number of candidate instruments.}
}
\value{
Character vector representing random IV model.
}
\description{
Function for generating a random IV model (combination of slab and spike priors).
}
\examples{
get_random_IV_model(5)
}
