% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{institutions_macroeconomic_growth_samples}
\alias{institutions_macroeconomic_growth_samples}
\title{A dataset containing results for reproducing Figure 6 in Section 6 of the UAI article.}
\format{
List containing posterior samples for the causal effect of rule-of-law
(institutions) on growth, as generated by two competing algorithms, MASSIVE and ivbma.
}
\usage{
institutions_macroeconomic_growth_samples
}
\description{
A dataset containing results for reproducing Figure 6 in Section 6 of the UAI article.
}
\keyword{datasets}
