% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JAMMR_MASSIVE_comparison}
\alias{JAMMR_MASSIVE_comparison}
\title{A dataset containing results for reproducing Figure 5 in Section 5 of the UAI article.}
\format{
A data frame containing evaluations of JAM-MR and MASSIVE in various
simulated scenarios.
\describe{
\item{Algorithm}{Name of algorithm, either JAM-MR or MASSIVE.}
\item{valid}{Integer number of valid instruments in the simulation.}
\item{beta}{Numeric value of beta (causal effect from X to Y) in the simulation.}
\item{rmse}{Root-mean-square error (RMSE) achieved by the algorithm.}
\item{rmse_se}{Standard error for RMSE metric across replications.}
\item{rmse_lower_ci}{Bootstrapped confidence interval 2.5\% quantile for RMSE across replications.}
\item{rmse_upper_ci}{Bootstrapped confidence interval 97.5\% quantile for RMSE across replications.}
\item{N:sigma}{Simulation configuration (number of observations):(noise magnitude).}
}
}
\usage{
JAMMR_MASSIVE_comparison
}
\description{
A dataset containing results for reproducing Figure 5 in Section 5 of the UAI article.
}
\keyword{datasets}
