% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers-data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Marker antibodies list}
\format{A data frame with 395 observations and 3 variables:
\describe{
 \item{marker}{}
 \item{label}{}
 \item{class}{}
 }}
\source{
\url{https://www.fluidigm.com/catalog?}
}
\usage{
markers
}
\description{
A dataset containing a comprehensive list with marker antibodies, which is
largely based on the Fluidigm catalog for CyTOF.
}
\keyword{datasets}
