% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE-package.R, R/MASSIVE_algorithm.R
\docType{package}
\name{MASSIVE}
\alias{MASSIVE}
\title{MASSIVE: Model Assessment and Stochastic Search for Instrumental Variable Estimation}
\usage{
MASSIVE(
  J,
  N,
  SS,
  sigma_G,
  sd_slab,
  sd_spike,
  max_iter = 1000,
  greedy_search = parallel_greedy_search,
  Laplace_approximation = safe_smart_LA_log
)
}
\arguments{
\item{J}{Integer number of candidate instruments}

\item{N}{Integer number of observations}

\item{SS}{Numeric moments matrix}

\item{sigma_G}{Numeric vector of instrument variances}

\item{sd_slab}{Numeric scale parameter of slab component}

\item{sd_spike}{Numeric scale parameter of spike component}

\item{max_iter}{Maximum number of stochastic search steps}

\item{greedy_search}{Greedy search function}

\item{Laplace_approximation}{Function computing the approximated model log-evidence}
}
\value{

}
\examples{
set.seed(2020)
J <- 5
N <- 1000
G <- matrix(rbinom(N * J, 2, 0.3), N, J)
G1 <- rbinom(N, 2, 0.3)
G2 <- rbinom(N, 2, 0.3)
U <- rnorm(N)
X <- G \%*\% rep(2, J) + U + rnorm(N)
Y <- G \%*\% rep(1, J) + U + X + rnorm(N)

Z <- cbind(1, G, X, Y)
SS <- t(Z) \%*\% Z / N
sigma_G <- apply(G, 2, sd)

samples <- MASSIVE::MASSIVE(J, N, SS, sigma_G, sd_slab = 1, sd_spike = 0.01)
plot(samples$betas)
}
