% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_log_prior}
\alias{neg_log_prior}
\title{Rcpp routine for computing the negative log-prior distribution of the MASSIVE model.}
\arguments{
\item{param_list}{List of parameter values in Rcpp format.}

\item{prior_sd}{List of prior standard deviations in Rcpp format.}
}
\value{
numeric value; negative log-prior value for the MASSIVE model.
}
\description{
Rcpp routine for computing the negative log-prior distribution of the MASSIVE model.
}
\keyword{internal}
