% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_search.R
\name{neighbor_IV_models}
\alias{neighbor_IV_models}
\title{Function for returning the neighboring IV model, i.e., those for which a
signel component changes from 0 (slab) to 1 (spike) or vice versa.}
\usage{
neighbor_IV_models(model, J)
}
\arguments{
\item{model}{Character vector describing IV model.}

\item{J}{Integer number of candidate instruments.}
}
\value{
List of IV models neighboring (one component changes) given IV model.
}
\description{
Function for returning the neighboring IV model, i.e., those for which a
signel component changes from 0 (slab) to 1 (spike) or vice versa.
}
\keyword{internal}
