% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_search.R
\name{parallel_greedy_search}
\alias{parallel_greedy_search}
\title{Routine to perform greedy search at the beginning of MASSIVE algorithm}
\usage{
parallel_greedy_search(
  J,
  N,
  SS,
  sigma_G,
  sd_slab = 1,
  sd_spike = 0.01,
  init_model = NULL,
  get_neighbors = smart_neighbor_models,
  LA_function = safe_smart_LA_log
)
}
\arguments{
\item{J}{}

\item{N}{}

\item{SS}{}

\item{sigma_G}{}

\item{sd_slab}{}

\item{sd_spike}{}

\item{init_model}{}

\item{get_neighbors}{}

\item{LA_function}{}
}
\value{

}
\description{
Routine to perform greedy search at the beginning of MASSIVE algorithm
}
