% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_list_to_vector}
\alias{parameter_list_to_vector}
\title{Function for converting list of parameters to vector format, often necessary
when running optimization routines.}
\usage{
parameter_list_to_vector(param_list)
}
\arguments{
\item{param_list}{List of MASSIVE model parameters.}
}
\value{
Numeric vector containing same parameters.
}
\description{
Function for converting list of parameters to vector format, often necessary
when running optimization routines.
}
\examples{
par <- random_Gaussian_parameters(10)
parameter_list_to_vector(par)
}
