% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_vector_to_list}
\alias{parameter_vector_to_list}
\title{Function for converting vector of parameters to list format.}
\usage{
parameter_vector_to_list(param_vector)
}
\arguments{
\item{param_vector}{Vector of MASSIVE model parameters.}
}
\value{
List containing same parameters.
}
\description{
Function for converting vector of parameters to list format.
}
