% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_comparison_samples}
\alias{prior_comparison_samples}
\title{A dataset containing results for reproducing Figure 4 in Section 5 of the UAI article.}
\format{
A list containing the posterior samples obtained with the three
priors under comparison together with the data simulation details.
\describe{
\item{Gaussian}{Posterior samples obtained with rstan when using Gaussian
prior for the MASSIVE model parameters.}
\item{Oracle}{Posterior samples obtained with rstan when using the Oracle
spike-and-slab prior for the MASSIVE model parameters.}
\item{MASSIVE}{Posterior samples output by MASSIVE algorithm.}
\item{simulation_details}{Details of the data and generating model in
the simulation.}
}
}
\usage{
prior_comparison_samples
}
\description{
A dataset containing results for reproducing Figure 4 in Section 5 of the UAI article.
}
\keyword{datasets}
