% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC3_sampler.R
\name{propose_neighbor_IV_model}
\alias{propose_neighbor_IV_model}
\title{Proposal function for generating a random neighbor of an IV model.}
\usage{
propose_neighbor_IV_model(IV_model, J)
}
\arguments{
\item{IV_model}{Character vector encoding IV model.}

\item{J}{Integer number of candidate instruments.}
}
\value{
Character vector encoding random neighbor of IV model.
}
\description{
Proposal function for generating a random neighbor of an IV model.
}
\keyword{internal}
