% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE_algorithm.R
\name{prune_model_list}
\alias{prune_model_list}
\title{Function to prune approximate IV model list returned by initial MASSIVE search.}
\usage{
prune_model_list(discovered_models, prob_threshold = 0.003)
}
\arguments{
\item{discovered_models}{List of discovered approximate IV models.}

\item{prob_threshold}{Numeric value indicating minimum probability threshold
that an IV model must have to avoid pruning}
}
\value{
List of pruned promising IV models.
}
\description{
Function to prune approximate IV model list returned by initial MASSIVE search.
}
\keyword{internal}
