% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_Gaussian_parameters}
\alias{random_Gaussian_parameters}
\title{Generate random Gaussian distributed parameters}
\usage{
random_Gaussian_parameters(J)
}
\arguments{
\item{J}{Integer number of instrumental variables.}
}
\value{
List of random parameters for the IV model generated from Gaussian distributions.
}
\description{
Generate random Gaussian distributed parameters
}
\examples{
random_Gaussian_parameters(10)
}
