% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCSFiles.R
\name{readFCSFiles}
\alias{readFCSFiles}
\title{Reads .fcs files}
\usage{
readFCSFiles(path = ".", FCSMeta, asinh5 = TRUE, downsampling)
}
\arguments{
\item{path}{A character string indicating the location of the .fcs files. The
default corresponds to the working directory}

\item{FCSMeta}{A list holding meta information on the .fcs files. The default
is set to \code{\link[massive]{readFCSMeta}}}

\item{asinh5}{Logical, should the data be transformed with the arc-sinh
cofactor 5}

\item{downsampling}{An integer that indicates the size to downsample to.
If not supplied, no downsampling will be performed.}
}
\description{
A wrapper function around \code{\link[flowCore]{read.FCS}} to read .fcs
files. It will only keep relevant channels and renames them according to
the `FCSMeta` parameter.
}
\examples{
FCSDir <- system.file("extdata", package="massive")
FCSData <- readFCSFiles(path=FCSDir)

}
\keyword{FCS,}
\keyword{data,}
\keyword{files}
\keyword{read,}
