% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace_approximation.R
\name{robust_safe_smart_LA_log}
\alias{robust_safe_smart_LA_log}
\title{Title}
\usage{
robust_safe_smart_LA_log(
  J,
  N,
  SS,
  sigma_G,
  prior_sd,
  post_fun = scaled_nl_posterior_log,
  gr_fun = scaled_nl_gradient_log,
  hess_fun = scaled_nl_hessian_log,
  opt_fun = robust_find_optimum,
  starting_points = "smart"
)
}
\arguments{
\item{J}{}

\item{N}{}

\item{SS}{}

\item{sigma_G}{}

\item{prior_sd}{}

\item{post_fun}{}

\item{gr_fun}{}

\item{hess_fun}{}

\item{opt_fun}{}

\item{starting_points}{}
}
\value{

}
\description{
Title
}
