% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE_algorithm.R
\name{sample_from_BMA_posterior}
\alias{sample_from_BMA_posterior}
\title{Sample from BMA posterior derived using the MASSIVE approach}
\usage{
sample_from_BMA_posterior(
  J,
  N,
  list_models,
  num_samples = 10000,
  log_sd = TRUE
)
}
\arguments{
\item{J}{Integer number of candidate instruments.}

\item{N}{Integer number of observations.}

\item{list_models}{List of IV models found by running MASSIVE}

\item{num_samples}{Integer number of samples to generate.}

\item{log_sd}{Flag indicating whether scale parameters are given in log-scale.}
}
\value{
List containing samples from BMA posterior.
}
\description{
Sample from BMA posterior derived using the MASSIVE approach
}
