% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE_model.R
\name{scaled_nl_gradient_log}
\alias{scaled_nl_gradient_log}
\title{Routine for computing the posterior gradient for the MASSIVE model}
\usage{
scaled_nl_gradient_log(J, N, SS, sigma_G, param_list, prior_sd, n = 2)
}
\arguments{
\item{J}{Integer number of genetic instrumental variables.}

\item{N}{Integer number of observations.}

\item{SS}{Numeric matrix containing first- and second-order statistics.}

\item{sigma_G}{Numeric vector of genetic IV standard deviations.}

\item{param_list}{List of IV model parameter values.}

\item{prior_sd}{List of standard deviations for the parameter Gaussian priors.}

\item{n}{Integer number of alleles (trials) for the binomial genetic variants.}
}
\value{
numeric vector; posterior gradient for the MASSIVE model.
}
\description{
Routine for computing the posterior gradient for the MASSIVE model
}
