% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE_model.R
\name{scaled_nl_posterior_log}
\alias{scaled_nl_posterior_log}
\title{Routines for computing the posterior distribution of the MASSIVE model}
\usage{
scaled_nl_posterior_log(J, N, SS, sigma_G, param_list, prior_sd, n = 2)
}
\description{
Routines for computing the posterior distribution of the MASSIVE model
}
