% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace_approximation.R
\name{smart_LA_log}
\alias{smart_LA_log}
\title{Routine for computing the Laplace approximation of the MASSIVE posterior}
\usage{
smart_LA_log(
  J,
  N,
  SS,
  sigma_G,
  prior_sd,
  post_fun = scaled_nl_posterior_log,
  gr_fun = scaled_nl_gradient_log,
  hess_fun = scaled_nl_hessian_log,
  opt_fun = robust_find_optimum,
  starting_points = "smart"
)
}
\arguments{
\item{J}{Integer number of candidate instruments}

\item{N}{Integer number of observations}

\item{SS}{Numeric moments matrix}

\item{sigma_G}{Numeric vector of instrument variances}

\item{prior_sd}{List containing prior hyperparameters}

\item{post_fun}{Function computing the negative log-posterior}

\item{gr_fun}{Function computing the negative log-posterior gradient}

\item{hess_fun}{Function computing the negative log-posterior Hessian}

\item{opt_fun}{Function for optimizing over the negative log-posterior}

\item{starting_points}{Character string describing the preset optimization starting points list}
}
\value{
List containing mixture of Laplace approximations at the discovered optima.
}
\description{
Routine for computing the Laplace approximation of the MASSIVE posterior
}
