% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_search.R
\name{stochastic_greedy_search}
\alias{stochastic_greedy_search}
\title{Routine to perform greedy search at the beginning of MASSIVE algorithm.
Here neighbors are examined in a random order until the first one that
improves the log-evidence is found and selected.}
\usage{
stochastic_greedy_search(
  J,
  N,
  SS,
  sigma_G = NULL,
  sd_slab = 1,
  sd_spike = 0.01,
  init_model = NULL,
  get_neighbors = IV_neighbor_models,
  LA_function = safe_smart_LA_log
)
}
\arguments{
\item{J}{Integer number of candidate instruments}

\item{N}{Integer number of observations}

\item{SS}{Numeric moments matrix}

\item{sigma_G}{Numeric vector of instrument variances}

\item{sd_slab}{Numeric scale parameter of slab component}

\item{sd_spike}{Numeric scale parameter of spike component}

\item{init_model}{Character string describing initial model}

\item{get_neighbors}{Function computing the negative log-posterior gradie}

\item{LA_function}{Function computing the approximated model log-evidence}

\item{tol}{Numeric tolerance stopping value for optimization}

\item{post_fun}{Function computing the negative log-posterior}
}
\value{
Character string indicating model found using greedy search
}
\description{
Routine to perform greedy search at the beginning of MASSIVE algorithm.
Here neighbors are examined in a random order until the first one that
improves the log-evidence is found and selected.
}
