% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet2j.R
\name{wet2j}
\alias{wet2j}
\title{Convert wet weight to approximate energy content}
\usage{
wet2j(wet)
}
\arguments{
\item{wet}{A vector of wet masses (in kg)}
}
\value{
The equivalent energy in Joules
}
\description{
This performs a simple conversion of wet weight to energy via Peters' approximate conversion.
}
\examples{
v <- rnorm(20, 10, 2)
v
wet2j(v)
}
\references{
Peters, R.H. (1983) The Ecological Implications of Body Size. Cambridge: Cambridge University Press (Cambridge Studies in Ecology), pp. 231–296. \url{https://doi.org/10.1017/CBO9780511608551}, Appendix Ia
}
\author{
Francis Windram
}
\concept{converter}
