% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry2wet.R
\name{dry2wet}
\alias{dry2wet}
\title{Convert (ash) dry weight to approximate wet weight}
\usage{
dry2wet(dry, fctr = 5)
}
\arguments{
\item{dry}{A vector of dry masses in kg}

\item{fctr}{a factor between 3 and 10 to use in the conversion (as Peters specifies a range of possible conversions).}
}
\value{
The equivalent wet mass values in kg
}
\description{
This performs a simple conversion of ash dry weight to wet weight via Peters' approximate conversion.
}
\examples{
v <- rnorm(20, 10, 2)
v
dry2wet(v)
}
\references{
Peters, R.H. (1983) The Ecological Implications of Body Size. Cambridge: Cambridge University Press (Cambridge Studies in Ecology), pp. 231–296. \url{https://doi.org/10.1017/CBO9780511608551}, Appendix Ia
}
\author{
Francis Windram
}
\concept{converter}
