% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC3_sampler.R
\name{find_causal_models}
\alias{find_causal_models}
\title{Routine to find the most likely causal models using a simple MH model stochastic search}
\usage{
find_causal_models(
  J,
  N,
  SS,
  sigma_G,
  sd_slab = 1,
  sd_spike = 0.01,
  stop_after = 1000,
  propose_model = smart_proposal,
  LA_function = safe_smart_LA_log,
  greedy_start = NULL,
  keep_greedy_approximations = FALSE
)
}
\arguments{
\item{J}{}

\item{N}{}

\item{SS}{}

\item{sd_slab}{}

\item{sd_spike}{}

\item{iter}{}

\item{init_model}{}
}
\value{

}
\description{
Routine to find the most likely causal models using a simple MH model stochastic search
}
