% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC3_sampler.R
\name{find_causal_models}
\alias{find_causal_models}
\title{Routine to find the most likely causal models using a simple Metropolis-
Hastings IV model stochastic search}
\usage{
find_causal_models(
  J,
  N,
  SS,
  sigma_G,
  sd_slab = 1,
  sd_spike = 0.01,
  max_iter = 1000,
  propose_model = propose_neighbor_IV_model,
  LA_function = safe_Laplace_approximation,
  greedy_start = NULL,
  keep_greedy_approximations = FALSE,
  ...
)
}
\arguments{
\item{J}{Integer number of candidate instruments.}

\item{N}{Integer number of observations.}

\item{SS}{Numeric matrix containing first- and second-order statistics.}

\item{sigma_G}{Numeric vector of genetic IV standard deviations.}

\item{sd_slab}{Numeric scale parameter of slab component.}

\item{sd_spike}{Numeric scale parameter of spike component.}

\item{max_iter}{Integer defining maximum number of MCMC iterations.}

\item{propose_model}{Proposal function indicating new IV model candidates.}

\item{LA_function}{Function for computing the IV model Laplace approximation.}

\item{greedy_start}{List returned by greedy search to be used as starting point.}

\item{keep_greedy_approximations}{Logical flag asking if greedy search history
should be stored and used in the new MCMC search.}

\item{...}{Extra arguments to pass to Laplace_approximation function.}
}
\value{
A list containing the list of models explored and their approximations,
the number of models explored, the total number of MCMC iterations, as well
as the acceptance rate.
}
\description{
Routine to find the most likely causal models using a simple Metropolis-
Hastings IV model stochastic search
}
\examples{
J <- 5 # number of instruments
N <- 1000 # number of samples
parameters <- random_Gaussian_parameters(J) 
EAF <- runif(J, 0.1, 0.9) # EAF random values
dat <- generate_data_MASSIVE_model(N, 2, EAF, parameters)
find_causal_models(J, N, dat$SS, binomial_sigma_G(dat$SS), 1, 0.01)
}
