% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{institutions_macroeconomic_growth_samples}
\alias{institutions_macroeconomic_growth_samples}
\title{A dataset containing results for reproducing Figure 16 in Section 5 of the UAI article.}
\format{
An object of class \code{list} of length 2.
}
\usage{
institutions_macroeconomic_growth_samples
}
\description{
A dataset containing results for reproducing Figure 16 in Section 5 of the UAI article.
}
\keyword{datasets}
