% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSIVE.R, R/MASSIVE_algorithm.R
\docType{package}
\name{MASSIVE}
\alias{MASSIVE}
\title{MASSIVE: Model Assessment and Stochastic Search for Instrumental Variable Estimation}
\usage{
MASSIVE(
  J,
  N,
  SS,
  sigma_G,
  sd_slab,
  sd_spike,
  max_iter = 1000,
  greedy_search = parallel_greedy_search,
  Laplace_approximation = safe_smart_LA_log
)
}
\arguments{
\item{J}{}

\item{N}{}

\item{SS}{}

\item{sd_slab}{}

\item{sd_spike}{}

\item{max_iter}{Maximum number of stochastic search steps}

\item{greedy_search}{Greedy search function}

\item{LA_function}{Laplace approximation function}
}
\value{

}
