% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_handlers.R
\name{random_Gaussian_parameters}
\alias{random_Gaussian_parameters}
\title{Generate random Gaussian distributed parameters}
\usage{
random_Gaussian_parameters(J, log_scale = FALSE)
}
\arguments{
\item{J}{Integer number of instrumental variables.}

\item{log_scale}{Logical flag indicating whether scale parameters sigma_X
and sigma_Y should be computed on the log-scale (normally distributed), or
on the normal scale (log-normally distributed).}
}
\value{
List of random parameters for the IV model generated from Gaussian distributions.
}
\description{
Generate random Gaussian distributed parameters
}
\examples{
random_Gaussian_parameters(10)
random_Gaussian_parameters(10, log_scale = TRUE)
}
