% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior_optima.R
\name{robust_find_optimum}
\alias{robust_find_optimum}
\title{Routine for (more robustly) finding a MASSIVE posterior local optimum}
\usage{
robust_find_optimum(
  J,
  N,
  SS,
  sigma_G,
  prior_sd,
  skappa_X,
  skappa_Y,
  tol = 1e-06,
  post_fun = scaled_nl_posterior_log,
  gr_fun = scaled_nl_gradient_log,
  hess_fun = scaled_nl_hessian_log
)
}
\arguments{
\item{J}{}

\item{N}{}

\item{SS}{}

\item{sigma_G}{}

\item{prior_sd}{}

\item{skappa_X}{}

\item{skappa_Y}{}

\item{tol}{}

\item{post_fun}{}

\item{gr_fun}{}

\item{hess_fun}{}
}
\value{

}
\description{
Routine for (more robustly) finding a MASSIVE posterior local optimum
}
