% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_search.R
\name{stochastic_greedy_search}
\alias{stochastic_greedy_search}
\title{Greedy search function in the IV model space.}
\usage{
stochastic_greedy_search(
  J,
  N,
  SS,
  sigma_G = NULL,
  sd_slab = 1,
  sd_spike = 0.01,
  init_model = NULL,
  get_neighbors = neighbor_IV_models,
  LA_function = safe_smart_LA_log,
  ...
)
}
\arguments{
\item{J}{Integer number of candidate instruments..}

\item{N}{Integer number of observations.}

\item{SS}{Numeric matrix containing first- and second-order statistics.}

\item{sigma_G}{Numeric vector of genetic IV standard deviations.}

\item{sd_slab}{Numeric scale parameter of slab component.}

\item{sd_spike}{Numeric scale parameter of spike component.}

\item{init_model}{Character vector describing starting IV model in search.}

\item{get_neighbors}{Function to get neighbor IV models at every step.}

\item{LA_function}{Function for computing the IV model Laplace approximation}

\item{...}{Extra arguments to pass to Laplace approximation function.}
}
\value{
A list containing the optimum found with greedy search, the list of
IV models visited and their approximations, and the number of visited models.
}
\description{
At each greedy step, we explore the IV models in the neighbor
set one by one. As soon as one has higher evidence than the current IV model,
we greedily move to the better neighbor.
}
\examples{
J <- 5 # number of instruments
N <- 1000 # number of samples
parameters <- random_Gaussian_parameters(J) 
EAF <- runif(J, 0.1, 0.9) # EAF random values
dat <- generate_data_MASSIVE_model(N, 2, EAF, parameters)
stochastic_greedy_search(J, N, dat$SS, binomial_sigma_G(dat$SS), 1, 0.01)
}
